requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
fastapi
uvicorn[standard]
python-dotenv
python-multipart
pypdf
sentence-transformers
faiss-cpu
groq
langchain-text-splitters
numpy